feat(KONFLUX-9835): When repo got moved, that means it was deleted
Some info from Inderpreet:

In the past (I think before GitLab 16.0), deleting a project was a much
more immediate and final action. The problem with this approach was the
risk of accidental data loss. Restoring a deleted project often required
a complex and time-consuming process. In later versions Gitlab added a
kind of grace-period to keep the project with different name for some
days so that user can recover it in case it was an accidental deletion.
I think how it works now is as below:

First Deletion: Navigate to your project's Settings > General >
Advanced. Under the "Delete project" section, click the Delete project
button. This will change the project's status to "pending deletion" and
rename the repository to allow you to immediately create a new project
with the same name.

Second Deletion: After the first step, the "Delete project" button will
become available again. Clicking it a second time and confirming will
permanently and immediately delete the project.

+++ pkg/clients/gitlab/git.go
@@ -258,6 +258,22 @@ func (gc *GitlabClient) GetCommitStatusConclusion(statusName, projectID, commitS
 // DeleteRepositoryIfExists deletes a GitLab repository if it exists.
 // Returns an error if the deletion fails except for project not being found (404).
 func (gc *GitlabClient) DeleteRepositoryIfExists(projectID string) error {
+	getProj, getResp, getErr := gc.client.Projects.GetProject(projectID, nil)
+	if getErr != nil {
+		if getResp != nil && getResp.StatusCode == http.StatusNotFound {
+			return nil
+		} else {
+			return fmt.Errorf("Error getting project %s: %v", projectID, getErr)
+		}
+	}
+	if getProj.PathWithNamespace != projectID && strings.Contains(getProj.PathWithNamespace, projectID + "-deleted-") {
+		// We asked for repo like "jhutar/nodejs-devfile-sample7-ocpp01v1-konflux-perfscale"
+		// and got "jhutar/nodejs-devfile-sample7-ocpp01v1-konflux-perfscale-deleted-138805"
+		// and that means repo was moved by being deleted for a first
+		// time, entering a grace period.
+		return nil
+	}
+
 	resp, err := gc.client.Projects.DeleteProject(projectID)
 
 	if err != nil {
@@ -272,7 +288,7 @@ func (gc *GitlabClient) DeleteRepositoryIfExists(projectID string) error {
 	}
 
 	err = utils.WaitUntilWithInterval(func() (done bool, err error) {
-		_, getResp, getErr := gc.client.Projects.GetProject(projectID, nil)
+		getProj, getResp, getErr := gc.client.Projects.GetProject(projectID, nil)
 		if getErr != nil {
 			if getResp != nil && getResp.StatusCode == http.StatusNotFound {
 				return true, nil
@@ -280,6 +296,11 @@ func (gc *GitlabClient) DeleteRepositoryIfExists(projectID string) error {
 				return false, getErr
 			}
 		}
+		if getProj.PathWithNamespace != projectID && strings.Contains(getProj.PathWithNamespace, projectID + "-deleted-") {
+			return true, nil
+		}
+
+		fmt.Printf("Repo %s still exists: %v\n", projectID, getResp)
 		return false, nil
 	}, time.Second * 10, time.Minute * 5)
